Improve link score calculation

+++ src/wifi/wfbng_link.h
@@ -73,7 +73,7 @@ public:
     /// Process a 802.11 frame.
     void handle_80211_frame(const Packet &packet);
 
-    std::array<float, ANTENNA_COUNT> get_link_score() const;
+    std::array<int, ANTENNA_COUNT> get_link_score() const;
 
     int get_packet_loss() const;
 
@@ -107,8 +107,8 @@ protected:
 #endif
 
     std::shared_ptr<SignalQualityCalculator> signal_quality_calculator;
-    std::array<float, ANTENNA_COUNT> link_score_ = {}; // Percentage
-    int packets_lost_ = 0;                              // Number over the last second
+    std::array<int, ANTENNA_COUNT> link_score_ = {}; // Percentage
+    int packets_lost_ = 0;                           // Number over the last second
 
 #ifndef _WIN32
     // --------------- Adaptive link
